@@ -5,7 +5,7 @@ author: Joshua de Guzman <contact@joshuadeguzman.net>
 homepage: https://github.com/joshuadeguzman/simple_switch
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
